requirements.txt
streamlit==1.25.0
tensorflow==2.12.0
numpy==1.26.0
pillow==10.1.0
